Mise en commun

+++ app/src/main/java/com/example/ihmproject/alertfactory/Accident.java
@@ -1,7 +1,7 @@
 package com.example.ihmproject.alertfactory;
 
 public class Accident extends Alert{
-    public Accident(int longitude, int latitude, String title, String description){
+    public Accident(double longitude, double latitude, String title, String description){
         super(longitude, latitude, title, description, 2);
     }
 }
